fix: add logs
